refactor : Input component 수정
