--- src/Components/Input/Input.js
@@ -1,39 +1,0 @@
-import React, { useState } from 'react'
-import styled from 'styled-components/macro'
-
-export default function Input({ setSendNumArr }) {
-  const [inputNum, setInputNum] = useState('')
-
-  const handleChange = (e) => {
-    setInputNum(e.target.value)
-  }
-
-  const handleBlur = () => {
-    setSendNumArr(inputNum.split(','))
-  }
-
-  return (
-    <div>
-      <InputBox value={inputNum} onChange={handleChange} onBlur={handleBlur} />
-    </div>
-  )
-}
-const InputBox = styled.input`
-  width: 400px;
-  box-sizing: border-box;
-  border: 3px solid #ccc;
-  border-radius: 30px;
-  font-size: 16px;
-  background-color: white;
-  background-image: url('https://ifh.cc/g/EDgulH.png');
-  background-size: 25px;
-  background-position: 13px 9px;
-  background-repeat: no-repeat;
-  padding: 12px 20px 12px 47px;
-  transition: all 0.45s ease 0s;
-  /* margin: 30px 0; */
-  &:focus {
-    outline: none;
-    border: 3px solid #0090b2;
-  }
-`
